Changed usize to be std::size_t

+++ include/statsdb/defs.hpp
@@ -16,7 +16,7 @@ using i16   = std::int16_t;
 using i32   = std::int32_t;
 using i64   = std::int64_t;
 
-using usize = std::uintmax_t;
+using usize = std::size_t;
 
 template <typename T> concept Formattable = requires(T t) {
         fmt::format("{}", t);
